new abstraction for matches
