+++ bancho/msgHandler.js
@@ -48,7 +48,10 @@ module.exports = {
 		if (message.match(commandRegex)) {
 			message = message.substring(1);
 			let args = message.split(" ");
+
+			if (args[0] == "!mp") return;
 			let command = commands[args[0]];
+			if (!command) return;
 			let options = args.splice(1, 1);
 			try {
 				if (messageType == 0 && command.dm) {
