+++ bancho/recovery.js
@@ -1,4 +1,5 @@
 const { prisma } = require("../prisma");
+const { MatchManager } = require("./MatchManager");
 module.exports.recover = async () => {
 	let matches = await prisma.match.findMany({
 		where: {
@@ -19,8 +20,8 @@ module.exports.recover = async () => {
 
 	for (const match of matches) {
 		if (match.state < 8 && match.state != 3) {
-			let manager = new Match(match.id, match.mp_link);
-			await manager.init();
+			let manager = new MatchManager(match.id, match.mp_link);
+			await manager.createMatch();
 		}
 	}
 };
